requirements.txt
django==1.8.17
gevent==1.2.2
leancloud-sdk==1.12.0